Separate method Check exception
