+++ src/main/java/domain/application/BlackJackApplication.java
@@ -6,7 +6,7 @@ import domain.validity.ValidityCheck;
 import domain.view.ViewInput;
 
 public class BlackJackApplication {
-
+	
 	public static void main(String[] args) {
 		Game blackJack = Game.getInstance();
 		UserRepository userRepository = blackJack.getUserRepository();
