+++ src/main/java/domain/view/ViewInput.java
