+++ src/test/java/domain/view/ViewTest.java
@@ -5,7 +5,7 @@ import domain.view.ViewInput;
 public class ViewTest {
 
 	public static void main(String[] args) {
-		ViewInput.getPlayerNames();
+		
 	}
 	
 }
